lazy-load location, admin, settings routes; remove unused deps
- Split /location, /admin, /settings routes into lazy-loaded chunks
  using lazyRouteComponent, reducing main bundle from 828kB to 645kB
- Fix date-fns import to use sub-path (date-fns/formatDistanceToNow)
- Remove unused zod dependency
- Move shadcn-ui to devDependencies
- Add bundle analysis files (stats.html, bundle-data.json) to gitignore
