+++ package.json
@@ -41,11 +41,9 @@
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.56.1",
     "react-leaflet": "^5.0.0",
-    "shadcn-ui": "^0.9.5",
     "steam-user": "^5.2.1",
     "tailwind-merge": "^3.2.0",
-    "vite": "^8.0.9",
-    "zod": "^3.24.3"
+    "vite": "^8.0.9"
   },
   "devDependencies": {
     "@tailwindcss/postcss": "^4",
@@ -58,6 +56,7 @@
     "tailwindcss": "^4",
     "tsx": "^4.21.0",
     "tw-animate-css": "^1.2.8",
-    "typescript": "^5"
+    "typescript": "^5",
+    "shadcn-ui": "^0.9.5"
   }
 }
